ReAct/agentic-reAct-customer-support/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.21.0
pydantic>=2.1.0
httpx>=0.23.0
python-dotenv>=1.0.0
openai>=1.0.0
pytest>=7.0.0